Move BitIterator into root of crate.
